Update User Permitted in Namespace Logic
With this change the "user permitted in namespace" logic now properly
indicates what namespaces a user is allowed to access and which they
can't (i.e. because the install doesn't have access to that namespace,
or the user themselves does not have permissions to access that
namespace).  Therefore, commands like 'pgo show namespace' now return
the proper results.