Create OCP 3.11 Specific Spec for Deployer Container
This commit creates a spec specifically for use by OpenShift 3.11
users when installing via the "kubectl" install method.  This spec
binds the 'pgo-deployer-sa' ServiceAccount to the 'cluster-admin'
role as needed to ensure the deployer can create any RBAC required
by the Operator.  This specifically works around a defect in OCP
3.11 in which the Operator is unable to create roles it does not
currently have (specifically due to a deficiency with the "escalate"
verb for roles).

Being that this is an OpenShift specific container, a few defaults
have been changed from the standard 'postgresql-operator.yaml'.
Specifically, "DISABLE_FSGROUPS" is set to "true" as required to
run within the "restricted" SCC on OpenShift, and "NAMESPACE_MODE"
has been set to "readonly" since "dynamic" is unable to work
properly in OCP 3.11 due to the same known "escalate" issue.