Use Secret for PGMONITOR_PASSWORD
A secret is now used to populate the PGMONITOR_PASSWORD env
var when Metrics are enabled for a cluster.  Also the
PGMONITOR_PASSWORD env var is now only set on the initial primary
for a cluster, and not on replicas, since the initialization logic'
triggered due to this env var in the PG HA container should only be
run on a primary.