Only expose Service ports that are used for a PostgreSQL instance
The Service template would check for the presence of optional ports
(e.g. pgBadger, collect) and expose those. However, this check is
performed by seeing if the value of the pgBadger / collect port is
present. The pgcluster CRD sets these values by default and, as such,
would always expose the port regardless if pgBadger/collect were
enabled.

This adjusts the logic to only include the value of the collect/pgBadger
port in the Service template if collect/pgBadger are actually enabled
on the CR entry.

Issue: #1611