Report accurate PVC capacity for replicas in `pgo show cluster`
The PVC capacity that was displayed as part of `pgo show cluster` was
pulling from the values set on the pgcluster CRD, which, for replicas,
may not be accurate if the PVC size was changed.

This adjusts the PVC capacity lookup formula to get that information
from the PVC object itself (similar to `pgo df`). Additionally, this
modifies the printed output to more clearly indicate which PVCs belong
to which Pods.

Issue: [ch8509]