pgBackRest URI Path Style at Cluster Create
This update allows the pgBackRest URI style to be set to
use 'path' style in place of the default 'host' style when
connecting to S3 and similar storage solutions. This is made
possible with the "--repo1-s3-uri-style=path" command option,
and can be configured during the initial installation of the
Crunchy PostgreSQL Operator, or overridden with the new "pgo
create cluster" flag, "--pgbackrest-s3-uri-style=path".