pgBackRest Verify TLS Option for S3 at Cluster Create
This update allows the pgBackRest Verify TLS option to be
set during cluster creation. A value of 'n' will disable
the TLS verification, with 'y' as the default which maintains
TLS verification. This is made possible with the
"--repo1-s3-verify-tls" command flag, and can be configured during
the initial installation of the Crunchy PostgreSQL Operator, or
overridden with the new "pgo create cluster" flag,
"--pgbackrest-s3-verify-tls=n".