Goodbye dep, hello go modules
An unintended side effect of #1576 and others is that the
directory structure is now conducive to using go modules. After
my latest wrangling with dep to try to do some routine upgrades,
I thought it prudent to attempt to make the switch. It turns
out that it was not too bad this go around, so now seems as likely
of a time to try to make the switch.

Issue: #1647