Remove 'bind' & 'escalate' Verb Requirements
This commit removes the requirement to assign the PostgreSQL Operator
ServiceAccount 'bind' and 'escalate' permissions in order to have the
PostgreSQL Operator reconcile RBAC (i.e. create, update and or delete
its own ServiceAccounts, Roles and RoleBindings) within any targeted
namespaces.  More specifically, the Operator is no longer concerned
with exactly how it has been assigned the permissions needed to
reconcile RBAC, and will now now always attempt to reconcile RBAC in
any targeted namespaces regardless of how any required permissions
were assigned.

For users that do not wish to have the Operator reconcile RBAC in target
namespaces, this capability can be disabled via the 'DisableReconcileRBAC'
setting in the 'pgo.yaml' configuration file.  When RBAC reconciliation
is disabled, all RBAC must be created manually by a Kuberentes
administrator (since the Operator will not attempt to do so itself).

In support of this change, the various installation methods supported
by the PostgreSQL Operator have been updated to enable RBAC reconciliation
by default.  Additionally, the various configuration settings,
descriptions, etc. for the "Dynamic RBAC" capability have been renamed to
"Reconcile RBAC" to better align with similar functionality and
terminology within Kubernetes (e.g. the 'kubectl auth reconcile' command,
which provides similar functionality).  The updated settings for each
installer are as follows (with each defaulting to 'true'):

- kubectl: RECONCILE_RBAC (previously DYNAMIC_RBAC)
- Ansible: reconcile_rbac (previously dynamic_rbac)
- Bash: PGO_RECONCILE_RBAC (previously PGO_DYNAMIC_RBAC)

When the various "Reconcile RBAC" installer settings are set to true
(as they are by default), the installer will configure the PostgreSQL
Operator with the proper permissions needed to reconcile RBAC based on the
Namespace Operating Mode configured for the installation.  For instance,
when a "dynamic" Namespace Operating Mode is configured, a ClusterRole
is installed that allows the Operator to reconcile RBAC in any namespace
(present or future) within the Kubernetes cluster.  When a "readonly" or
"disabled" Namespace Operating Mode is configured, the Roles are installed
in each target namespace as needed to manage RBAC in those specific
namespaces (therefore, while in "readonly" or "disabled" modes the RBAC
required by the Operator to reconcile RBAC in any new namespaces must be
manually created by an administrator).

In support of the above changes, the 'Namespace' section of the
documentation has been updated, along with the applicable settings in the
various installer and configuration documentation.  This includes defining
RBAC required to allow the Operator to successfully reconcile RBAC in
targeted namespaces.  Additionally, this commit removes all remaining
wildcards ('*') from PostgreSQL Operator Role definitions.

And finally, this commit includes various refactoring for the namespace
controller and the associated Operator Controller Manager, specifically to
reduce duplication of various RBAC logic, improve error handling and
logging, and overall simplify and better structure/organize these portions
of the Operator code.