Allow support for "authentication" block in custom config
As this is a block that is generally managed by the PostgreSQL
Operator itself, access to customizing this block was not initially
allowed as "bad things could happen." However, there are certain
settings within this block (e.g. sslmode) that one may want to
modify, and we want to avoid the tedium of adding all of these
supported options into the Operator itself, at least at this
time.

As such, customizing the "authentication" block within the
"local" options for modifying a PostgreSQL cluster is now
supported, with the general warning that modifying certain
attributes (username, password) within that block can have
undesired consequences.

Issue: [ch8468]