Enable replication connections to use certificate-based authentication
Certificate-based authentication[1] is a powerful PostgreSQL feature
that allows for a PostgreSQL client to authenticate using a TLS
certificate. While there are a variety of permutations for this can be
set up, we can at least create a standardized way for enabling the
replication connection to authenticate with a certificate, as we do have
a known certificate authority.

This introduces the `--replication-tls-secret` flag on the
`pgo create cluster` command, which, if specified and if the
prerequisites are specified (`--server-tls-secret` and
`--server-ca-secret`), then the replication account ("primaryuser") is
configured to use certificate-based authentication. Combine with
`--tls-only` for powerful results.

Note that the common name (CN) on the certificate MUST be "primaryuser",
otherwise one must specify a mapping in a `pg_ident` configuration block
to map to "primary" user.

When mounted to the container, the connection sslmode that the
replication user uses is set to "verify-ca" by default. We can make that
guarantee based on the certificate authority that is being mounted.
Using "verify-full" would cause the Operator to make assumptions about
the cluster that we cannot make, and as such a custom `pg_ident`
configuration block is needed for that.

Issue: [ch8387]

[1] https://www.postgresql.org/docs/current/auth-cert.html