Pin Ansible version to 2.9.0+
This update standardizes the Ansible version across the different deployer
images and the documentation. The docs have been updated so that Ansible 2.9.0+
is a prerequisites for the Ansible installer (documented under other install
methods). The gcp-marketplace deployer image has been updated from Ansible 2.8.*
to 2.9.* and the pgo-deployer image has been pinned to 2.9.*.

The image specific environment variables are moved to the Dockerfile. The HOME
and ANSIBLE_CONFIG variables are currently set in the pgo-deploy script. This
causes issues when trying to use ansible without running the script or manually
setting these variables.