Ansible conversion from ini to yaml
Changes include updating the Ansible install to use a combination of a yaml
inventory file and values file in place of a single ini inventory file. The
inventory.yaml will include information that is either required by or only used
by Ansible. This includes host information and variables that are used to
connect to a Kubernetes or OpenShift cluster. The values.yaml file, which is now
included the the $PGOROOT/installers/ansible dir, will contain all configuration
options for the PostgreSQL Operator. This file should be at the path defined by
config_path in the inventory.yaml. The default for this path is
$PGOROOT/installers/ansible/values.yaml but can be updated in the inventory.yaml
file or at the command line.