Updates pgo-deployer to use configmap
The kubectl install method is now updated to use a ConfigMap to pass in this
values.yaml file instead of setting environment variables in the deployer image.
The postgres-operator.yml and postgres-operator-311.yml will create the required
ConfigMap and ensure that it is mounted as a volume to the pgo-deploy job. In
order to configure the PostgreSQL Operator deployment you will need to update
the values.yaml section that is defined in your manifest file. This file will
then be used to run the Ansible installer in the pgo-deployer image. The
DEPLOY_ACTION is still used to define which Ansible tag will be used and should
be defined in the postgres-operator.yml (or postgres-operator-311.yml).

The pgo-deployer image has been updated to work with the updates from the
Ansible install and now makes use of the ConfigMap defined in the kubectl job
manifest. The image now uses an updated ansible-playbook command that overwrites
the config_path and enables kubernetes_in_cluster. This set of changes updates
is an attempt to standardize the input for the pgo-deployer. The image now takes
a generic yaml configuration file (values.yaml) that can be use across different
install methods without having to maintain multiple files.