Use a clientset composed of upstream and local
We often pass a versioned.Interface alongside a kubernetes.Interface and
distinguishing between the two can be a pain. This commit introduces an
interface and type in kubeapi to satisfy both at once.

All instances of *rest.RESTClient are now interfaces. `pgo-scheduler` no
longer depends on internal/apiserver.

Issue: [ch8733]