Create Cluster from Backup
This change allows a new PostgreSQL cluster to be bootstrapped using an
existing data source, specifically an existing pgBackRest repository.  The
repository that is used to bootstrap the cluster can either be a
repository from a former PostgreSQL cluster that no longer exists
(assuming backups and the pgBackRest repository secret were retained by
specifying the '--keep-backups' when deleting it), or from a cluster
that is currently running.

When bootstrapping from a previous cluster that no longer exists, the
Operator will temporarily deploy a pgBackRest repository host for that
repository for purposes of bootstrapping the new cluster.  The new
cluster being created will then connect to that repository using SSH
to perform the restore needed to initially populate its PGDATA
directory.  On the other hand, when bootstrapping from a cluster that
is currently running, the new cluster being created will simply connect
(also via SSH) to the pgBackRest repository host that is currently running
for that cluster.

The actual bootstrapping of the new cluster, i.e. the pgBackRest restore
that is executed to populate the PGDATA directory for the new cluster,
is done via a Kubernetes Job.  This Job will mount the secrets needed
to access the appropriate pgBackRest repository host for the cluster being
restored from and execute the 'pgbackrest restore' command.  Once the Job
has completed successfully (i.e. the 'pgbackrest restore' command was
successful and any recovery completed), any resources created (e.g. a
repository host) created for the bootstrap are removed, and cluster
initialization (stanza-creation, initial backup, etc.) proceeds as usual.
The only difference is that the cluster will start with the PGDATA content
populated when the bootstrap Job was run, instead of initializing from
scratch using 'initdb'.

Two new flags have been added to the 'pgo create cluster' command to
specify an existing cluster to bootstrap from:

- restore-from: the name of the cluster (past or present) to restore
from when bootstrapping the cluster.
- restore-opts: the CLI options that should be passed to pgBackRest when
the 'pgbackrest restore' command is run (e.g. the '--type' and 'target'
flags if performing a PITR)

Only a single cluster can bootstrap from another cluster (running or not)
at any given time.  Additionally, the Operator will not allow a client
to bootstrap from an existing cluster if it is not in the proper status
(e.g. if it is in a 'shutdown' or 'restoring' status).