Adds metrics support for helm and pgo-deployer
Metrics support was an undocumented feature of the pgo-deployer. This commit
adds documentation around how you can use the `kubectl` job manifest and
`pgo-deployer` to install and uninstall the metrics infrastructure.

The deployment options in the new yaml files are updated so that the boolean
values are treated as bool instead of string.