Silence Disable Autofail ConfigMap Errors
With this commit, certain "ConfigMap not found" errors seen during
cluster initialization have been silenced.  Specifically, the function
call to startup a cluster within the pgcluster "onUpdate" controller
logic is now only executed when the cluster is in a Shutdown status,
preventing the startup logic (which includes an attempt to modify
the DCS ConfigMap to enable autofailover) from executing during cluster
initialization (and therefore possibly before the DCS ConfigMap
exists).

Additionally, an error message in the in the StartupCluster function
has been updated to indicate that the error occurred while attempting
to start the cluster (and not when shutting it down).