Add upgrade support for updated PostGIS container tag
With the update to the tagging of the Crunchy Postgres GIS HA
container, the Postgres Operator automated upgrade has been
updated to allow for this new tag during the upgrade process.

To upgrade a PostGIS enabled pgcluster, a new flag,
'--post-gis-image-tag' will be required to specify the specific
container tag value, as in the command below:

pgo upgrade --post-gis-image-tag=centos7-11.8-2.5-4.4.0 mygiscluster

Where "11.8" is the PostgreSQL version, "2.5" is the PostGIS version
and "4.4.0" is the PostgreSQL Operator version.