PostGIS enabled cluster creation
Due to the new tag value on the Crunchy Postgres GIS HA
container, the specific tag now must be supplied when
created a PostGIS enabled pgcluster. An example command is below:

pgo create cluster hagiscluster \
--ccp-image=crunchy-postgres-gis-ha \
--ccp-image-tag=centos7-11.8-2.5-4.4.0

Where "11.8" is the PostgreSQL version, "2.5" is the PostGIS version
and "4.4.0" is the PostgreSQL Operator version.