Fix for panic if config syncs when Pod unavailable
There is a potential race condition in which the PostgreSQL HA
configuration attempts to sync with a Pod, but the Pod is
currently unavailable. This causes the PostgreSQL Operator to
crash and potentially halt an operation (e.g. this was discovered
during an upgrade).

If no Pods matching the selector are found, we abort the HA
ConfigMap synchronization.

Issue: #1682