Updates client-setup.sh to handle an existing pgo file (#1708)
Currently the script will throw an error if the pgo binary file exists. This
change updates the script to overwrite the original file and download the new
binary in its place. In this case a message is printed telling the user that the
binary is being updated. Considering the binary is downloaded into a namespaced
directory overwriting the file is considered a safe option.