Rewrite docs to promote "pgo create cluster --restore-from"
"pgo create cluster --restore-from" is equivalent to "pgo clone"
but is much more robust, as it provides the ability to use the
full set of "pgo create cluster" options. As such, techniques
that use "pgo clone" are rewritten to showcase
"pgo create cluster --restore-from".

This also makes it much clearer how the request of #1305 can
be leveraged, as both the architecture of the feature and several
use cases are further fleshed out.