Remove Status & Pod Check from Pod Controller
Logic to return from the onUpdate() function of the Pod controller if
the cluster is not in an initialized status, or if the Pod is not a
PostgreSQL Pod (i.e. a primary or replica), has been removed.  This
is because the same logic is built into the function calls below it,
i.e. functions isPromotedPostgresPod() and isPromotedStandby()
both verify that the Pod is a PostgreSQL Pod, while the subsequent
handler functions check the current status of the pgcluster as
needed.