Master Config Update
This update adds a master config that is used for the Ansible, Helm, and Kubectl
installers. The config_sync.sh script is added to the hack directory, this
script will use the ansible/values.yaml file as the master config.

The values.yaml for each installer is updated to exactly match the master
config. This is a small change that includes adding a link to the docs and
syncing the pgo_client options.

The master config and the helm installer values.yaml will be out of sync until
the installers are updated with coming metrics changes. The master config still
contains metrics options but those options were removed from the helm
values.yaml.