Update bash installer to use json configs from ansible
Bash and Ansible installers are currently pulling json templates from two
different places. `$PGOROOT/conf` and
`$PGOROOT/installers/ansible/roles/pgo-operator/files` both contain the same
json templates. By setting both installers to pull the templates from the same
directory we can simplify maintenance when changing these files.

We are standardizing on Ansible as the main installer that is a base for each
other installer. This commit includes changes to make the Bash installer use the
Ansible config directory.

Bash is now setup to use an environment variable to determine the config
directory. This environment varible is updated to point to the Ansible config
directory.

The `pgo.yaml`, `server.crt`, and `server.key` files will still exist in
`$PGOROOT/conf` because the are not in the ansible config directory. Ansible
generates these files based on templates instead of having the files defined.