Add runAsUser '2' when fsGroup is Enabled
When fsGroup is enabled for non-OpenShift deployments, the runAsUser
is now also set to '2' in the securityContext for the Grafana and
Prometheus Deployment specs.  This ensures that the upstream Grafana
and Prometheus containers are able to access  Grafana and Prometheus
data from previous versions of the PostgreSQL Operator metrics
infrastructure (which utilized the 'crunchy-grafana' and
'crunchy-prometheus' containers that ran using UID 2).

Additionally, the Grafana data volume mount has been updated and
configured to reflect the data path utilized by 'crunchy-grafana' in
previous versions of the metrics infrastructure.  This ensures a
previous Grafana volume will be properly mounted when upgrading to the
new metrics infrastructure.