An Monitoring Intro to Installation Section
The "PostgreSQL Operator Monitoring" installation page now provides an
overview of PostgreSQL Operator Monitoring infrastructure.  This
includes an overview of the capabilities and tools included with the
monitoring solution, as well as how the installation can be customized
and deployed alongside the Operator according to various use cases and
needs.