Default URI Style for S3 Bootstrap
The proper default S3 URI style is now set when generating the
pgBackRest S3 environment variables for a bootstrap Job (i.e. a
Job created to bootstrap a cluster from an existing backup, such
as when using the '--restore-from' flag with the 'pgo' client).
Specifically, if an empty string is detected in the 's3-uri-style'
annotation for the pgBackRest repo secret for the cluster being
bootstrapped/restored from, a proper default value of 'host' is now
set.  This ensures there is never an empty value for the
PGBACKREST_REPO1_S3_URI_STYLE environment variable in a bootstrap
Pod, which can lead to problems when running various pgBackRest
commands.

Additionally, tests have been created to verify the output of the
GetPgbackrestBootstrapS3EnvVars function that is utilized to populate
the proper S3 environment variables for a bootstrap Job (specifically
using the pgBackRest repo secret from the cluster being bootstrapped
from).  This includes ensuring the proper default URI style is set
when the 's3-uri-style' annotation is empty, as well as ensuring
all annotation values are properly reflected in the final S3
environment variable output.  In support of these tests, a
NewFakePGOClient function has also been created in the 'fake' package
(underneath 'kubeapi').  This is needed to create a fake client
containing the proper resources (i.e. a mock 'pgo-config' ConfigMap)
as need to initialize the Operator and therefore the various templates
needed to properly call the GetPgbackrestBootstrapS3EnvVars function
for these tests (PGOROOT must be properly set in the environment when
running tests so that the proper default template files and pgo.yaml
file can be found as needed to create the mock 'pgo-config' ConfigMap).

Issue: [ch8915]