Custom Annotations for Deployments
This allows one to add custom annotations to the managed
deployments, including the PostgreSQL Operator, pgBackRest, and
pgBouncer. Annotations can be applied to all of the Deployments,
or tuned to be applied to the individual Deployment groups.

This also introduces several flags to the PostgreSQL Operator
client that help with the management of the annotations. These
flags are available on `pgo create cluster` and
`pgo update cluster`:

- --annotation - apply annotations on all managed Deployments
- --annotation-postgres - " managed Postgres Deployments
- --annotation-pgbackrest - " managed pgBackRest Deployments
- --annotation-pgbouncer - " managed pgBouncer Deployments

All the flags can be used as a list.

To add an annotation, one follows the format:

	`--annotation=key=value`

To remove an annotaiton, one follows the format:

	`--annotation=key-`

Issue: [ch7938]