Remove `pgo load`
The functionality had not worked in awhile, the interface
could present challenges to use, and there are other was
that data can presently be loaded.

This gives us an opportunity to explore future implementations
on how we can permit data loading into PostgreSQL, but likely
this is application specific.

Issue: [ch9061]