Add Explicit 'else' to Jinja Inline If Expression
Per the Jinja documentation, the 'else' for an inline If expression
should default to an undefined object if not explicitly specified:

https://jinja.palletsprojects.com/en/2.11.x/templates/#if-expression

However, with certain versions/combinations of Ansible and Python,
this default is not occurring, leading to errors when the "Wait for
Metrics to Finish Deploying" task is run during a metrics installation.

This commit therefore adds an explicit 'else' condition to the inline
If expression within this task, ensuring maximum compatibility and
proper behavior with various combinations/versions of Ansible and
Python.