Incorporate dashboards of pgMonitor 4.4-RC4
This pgMonitor version has dashboards designed for the Kubernetes labels
and metrics provided by our exporter container. The dashboards are
organized into their own directory, so we no longer need to pick and
choose which to import.

The dashboards are exported from Grafana for external use, so we must
select/replace the Prometheus datasource when populating the ConfigMap
(which in turn *provisions* the dashboards at startup).

 - https://grafana.com/docs/grafana/latest/reference/export_import/
 - https://grafana.com/docs/grafana/latest/administration/provisioning/

Issue: [ch8765]