Python Auto-Detect & Install openssl in Deployer
The 'ansible_python_interpreter' setting has been removed from the
inventory file for the PostgreSQL Operator Ansible installer, which
allows Ansible to auto-detect Python within the environment it is
running.  This ensures the proper detection and use of Python by
Ansible within the deployer container across the various versions
of UBI and CentOS supported by the PostgreSQL Operator.

Additionally, the 'openssl' package is now explicitly installed in the
UBI8 deployer container, as needed to generate various certificates
when the Ansible installer is run run within the container.