Remove Obsolete Exporter Function Call
This commit removes an obsolete reference to the
'set_exporter_pg_credentials' function from the Crunchy Postgres
Exporter container's start.sh script. This function was previously
needed to read the credential values from the containers volume, but
these values are now set in the deployment. While the function itself was
previously removed, this reference was missed, resulting in a misleading
error in the logs.