Fix --target Option Detection for Restores
A regex is now used to determine whether or not a "--target" flag
already exists in the options provided for a pgBackRest restore.
This avoids conflicts with other options containing the string "target"
(e.g. "--target-action"), therefore ensuring the PITR target is
properly injected into the pgBackRest options as needed.

Issue: [ch9162]