Shorter pgrestore Job Names
The name of the PVC being utilized for a pgrestore Job is now no longer
included in the Job name.  This ensures the 63 character limit for
Kubernetes Job naming is not easily exceeded when creating pgrestore
Jobs, i.e. for clusters with names longer than simply a few characters.

Issue: [ch9163]