Move pgBouncer configuration to use ConfigMaps
This moves nonsensitive pgBouncer configuration to be located
in a ConfigMap that sits alongside an Operator managed cluser
in the same namespace. Specifically, this moves two entries:

- pgbouncer.ini - controls the general configuration
- pg_hba.conf - controls how one can connect to pgBouncer

This presently does not handle any automatic updates if the
ConfigMap entry is changed, but rather is just a step to move
the nonsensitive configurations out of the Secret.

Issue: [ch568]