Update ansible values.yaml pgo_client_version
A github url is used to download the pgo client when installing with ansible
locally. In the main config update the ansible values.yaml file was updated so
that the "v" in "v4.4.1" was defined in the download url. This changes updates
the values.yaml to account for this change.