Prevent Panic if PG Deployments Scaled Down
If the primary and all replica Deployments are scaled to zero outside
of the context of the PostgreSQL Operator (e.g. by using
'kubectl scale deployment --replicas=0' instead of
'pgo update cluster --shutdown') and the associated pgcluster for the
Deployment(s) being scaled is not manually set to the proper "shutdown"
status, then the PostgreSQL Operator could panic as it continues to
attempt to synchronize Patroni/PostgreSQL configuration for the PG
cluster (specifically because no pods are found for the cluster).  This
commit protects against a panic in this scenario by simply logging an
error when no Pods are found during an attempt to sync configuration.

Issue: [ch9278]