Update annotation update behavior
The existing annotation update logic is unable to maintain
existing 'global' annotations when a 'specific' (postgres, pgbackrest or
pgbouncer) deployment annotation is updated (e.g. the Postgres only
annotation is updated, but replaces any existing 'global' annotations
that were made previously) and vice versa.

To remedy this, if either 'global' or 'specific' deployment annotations
are updated, the relevant map is updated to include all existing 'global'
or 'specific' annotations from the pgcluster CRD. 'Global' annotations with
the same key value of a 'specific' annotation will be overwritten, so that
the more specific annotation is authoritative.